third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-style-item.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This tests indexing outside the range of the style object.");

var element = document.createElement('div');
element.style.color = 'red';
document.documentElement.appendChild(element);
var style = element.style;

shouldEvaluateTo('style.length', 1);
shouldBeEqualToString('style[0]', 'color');
shouldBeUndefined('style[1]');
shouldBeUndefined('style[-1]')

document.documentElement.removeChild(element);
</script>
</body>
</html>
