third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-style-item-expected.txt

This tests indexing outside the range of the style object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.length is 1
PASS style[0] is "color"
PASS style[1] is undefined.
PASS style[-1] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
